requirements.txt
pyodbc==4.0.39
mysql-connector-python==8.0.33
schedule==1.2.0
python-dotenv==1.0.0
pytest==7.4.0
pylint==2.17.5
black==23.7.0
